+++ subprojects/kotlin-dsl/src/main/kotlin/org/gradle/kotlin/dsl/KotlinSettingsScript.kt
@@ -49,7 +49,7 @@ import org.gradle.internal.time.Clock
 
 import org.gradle.kotlin.dsl.resolver.KotlinBuildScriptDependenciesResolver
 import org.gradle.kotlin.dsl.support.KotlinScriptHost
-import org.gradle.kotlin.dsl.support.decorators.SettingsDecorator
+import org.gradle.kotlin.dsl.support.delegates.SettingsDelegate
 import org.gradle.kotlin.dsl.support.get
 import org.gradle.kotlin.dsl.support.internalError
 import org.gradle.kotlin.dsl.support.serviceOf
@@ -108,7 +108,7 @@ abstract class KotlinSettingsScript(
  */
 abstract class SettingsScriptApi(
     override val delegate: Settings
-) : SettingsDecorator() {
+) : SettingsDelegate() {
 
     protected
     abstract val fileOperations: FileOperations
