+++ subprojects/kotlin-dsl/src/main/kotlin/org/gradle/kotlin/dsl/support/KotlinPluginsBlock.kt
@@ -20,7 +20,7 @@ import org.gradle.api.Project
 import org.gradle.api.initialization.dsl.ScriptHandler
 
 import org.gradle.kotlin.dsl.ScriptHandlerScope
-import org.gradle.kotlin.dsl.support.decorators.ProjectDecorator
+import org.gradle.kotlin.dsl.support.delegates.ProjectDelegate
 
 import org.gradle.plugin.use.PluginDependenciesSpec
 
@@ -42,7 +42,7 @@ abstract class KotlinPluginsBlock(val pluginDependencies: PluginDependenciesSpec
 abstract class KotlinBuildscriptAndPluginsBlock(
     private val host: KotlinScriptHost<Project>,
     val pluginDependencies: PluginDependenciesSpec
-) : ProjectDecorator() {
+) : ProjectDelegate() {
 
     override val delegate: Project
         get() = host.target
