+++ subprojects/kotlin-dsl/src/main/kotlin/org/gradle/kotlin/dsl/support/delegates/ArtifactHandlerDelegate.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.kotlin.dsl.support.decorators
+package org.gradle.kotlin.dsl.support.delegates
 
 import groovy.lang.Closure
 
@@ -27,9 +27,9 @@ import org.gradle.api.artifacts.dsl.ArtifactHandler
 /**
  * Facilitates the implementation of the [ArtifactHandler] interface by delegation via subclassing.
  *
- * See [GradleDecorator] for details why this is currently necessary.
+ * See [GradleDelegate] for details why this is currently necessary.
  */
-abstract class ArtifactHandlerDecorator : ArtifactHandler {
+abstract class ArtifactHandlerDelegate : ArtifactHandler {
 
     internal
     abstract val delegate: ArtifactHandler
